requirements.txt
clickhouse-connect>=0.6.0
python-binance>=1.0.0
